Remove colour support
This stops the homekit app from showing a bunch of horrible colours as
the only choices on the color swatches, and instead shows different
whites from warm to cool.

I also pulled down the max kelvin so the min and max values are
fully supported on my LIFX original bulbs and my new LIFX mini.